+++ packages/buildpacks/test/unit/commands/buildpacks/clear.unit.test.ts
@@ -1,8 +1,8 @@
 import {expect, test} from '@oclif/test'
 import * as nock from 'nock'
 
-import {BuildpackInstallationsStub as Stubber} from '../../helpers/buildpack-installations-stub'
-import {unwrap} from '../../unwrap'
+import {BuildpackInstallationsStub as Stubber} from '../../../helpers/buildpack-installations-stub'
+import {unwrap} from '../../../unwrap'
 nock.disableNetConnect()
 
 describe('buildpacks:clear', () => {
