+++ packages/buildpacks/test/unit/commands/buildpacks/set.unit.test.ts
@@ -1,7 +1,7 @@
 import {expect, test} from '@oclif/test'
 import * as nock from 'nock'
 
-import {BuildpackInstallationsStub as Stubber} from '../../helpers/buildpack-installations-stub'
+import {BuildpackInstallationsStub as Stubber} from '../../../helpers/buildpack-installations-stub'
 nock.disableNetConnect()
 
 describe('buildpacks:set', () => {
